// src/types.ts
export type InventoryStatus = 'AMAN' | 'MENIPIS' | 'HABIS';

export interface InventoryItem {
  id: string;
  store_id: string;
  category_id: string;
  kode_barang: string;
  nama_barang: string;
  model_iphone: string;
  harga_beli: number;
  harga_jual: number;
  stok: number;
  stok_minimum: number;
  satuan: string; // pcs, set, unit, roll, botol
  keterangan: string;
  status: InventoryStatus;
  created_at: string;
  updated_at: string;
}